chore: renamed routes to routers
